Add restore_state support for sensors
A lot of sensors (i.e. the Xiaomi temperature sensors) send state
updates only if the data has changed. There is no way to poll the
current state. Thus, after a restart of home assistant, it can take
hours or even days until a value is available.

Use the restore_state (together with the recorder component) to set
the state initially to the last known value.